src/data_processing/requirements.txt
google-cloud-storage==2.7.0
flask>=2.2
pandas==1.5.3
numpy==1.23.3
